fixing logic of best time
was working that system would check what's most common in days column
then check for most common in timeframe column it was to easy for the case
of it to select just a timeframe that appears more times then the most times paired
with the specific best day.

Now it checks the days column for what show's up the most then takes that number to find
what shows up the most in timeslots column. Also cleaned up more of the range to slot in terms
of timeframe